centered quill and added two components
